Moving signal functions from deneb-common to the main app crate
