+++ Cargo.toml
@@ -18,6 +18,7 @@ deneb-common = { path = "deneb-common" }
 deneb-core = { path = "deneb-core" }
 failure = "^0.1.0"
 log = "^0.4.0"
+log4rs = "^0.8.0"
 nix = "^0.9.0"
 structopt = "0.1.0"
 structopt-derive = "0.1.0"
