+++ deneb-common/Cargo.toml
@@ -6,6 +6,5 @@ authors = ["Radu Popescu <mail@radupopescu.net>"]
 [dependencies]
 failure = "^0.1.0"
 log = "^0.4.0"
-log4rs = "^0.8.0"
 nix = "^0.9.0"
 time = "^0.1.0"
