+++ deneb-common/src/lib.rs
@@ -1,10 +1,7 @@
 #[macro_use] extern crate failure;
 #[macro_use] extern crate log;
-extern crate log4rs;
 extern crate nix;
 extern crate time;
 
-
 pub mod errors;
-pub mod logging;
 pub mod util;
