+++ src/lib.rs
@@ -3,12 +3,10 @@
 extern crate failure;
 #[cfg(feature = "fuse")] extern crate fuse;
 extern crate log;
+extern crate log4rs;
 extern crate nix;
-#[cfg(test)] extern crate quickcheck;
-#[cfg(test)] extern crate rand;
 extern crate structopt;
 #[macro_use] extern crate structopt_derive;
-#[cfg(test)] extern crate tempdir;
 extern crate time;
 
 // Crates from the workspace
@@ -16,4 +14,6 @@ extern crate deneb_common;
 extern crate deneb_core;
 
 pub mod fe;
+pub mod logging;
 pub mod params;
+pub mod util;
